#297 resolve imports
