+++ allensdk/mouse_connectivity/grid/subimage/base_subimage.py
