+++ requirements.txt
@@ -16,3 +16,4 @@ simpleitk
 tables
 argschema
 marshmallow>=3.0.0b20
+glymur
